refactor(cli): centralize slug fuzzy match and provider validation
- Add slugScore/closestSlug helper used by check and show commands
- Import VALID_PROVIDERS from @ossperks/core in check command

Made-with: Cursor
